Moved helper function into seperate file
